trying to fix the sid...
Apparently the bug is in the extra 'jsr' added in the sid
code in order to save the sid status in "external" variables.
Storing only the $d404-$d406 seems to work ok.

Also, I'm not sure why this bug was cannot be reproduced in
the version with the hires logo...?!?